refactor: move dashboardHeaderMenuItems
