+++ frontend/dashboard/types/HeaderMenuItem.ts
@@ -1,0 +1,7 @@
+import type { HeaderMenuItemKey } from '../enums/HeaderMenuItemKey';
+
+export type HeaderMenuItem = {
+  key: HeaderMenuItemKey;
+  link: string;
+  name: string;
+};
