+++ frontend/dashboard/utils/headerUtils/index.ts
@@ -1,0 +1,1 @@
+export { dashboardHeaderMenuItems } from './headerUtils';
